- Removed all set/remove restrictions for the LIGHT flag. - Made it so that connections of type INBOUND are not logged/logwalled     when disconnected. Since this are handled in MUF, it is assumed that     the MUF code will handle any logging needed. - Changed all FLOAT precision to DOUBLE instead. [FB6 change] - Made it so that the 'Program not compilable' message reports the dbref. - Made it so that the notify_descriptor* prims don't abort when trying to     notify to an invalid descriptor. They simply don't do anything in     such a case. - Added MUF EVENT pid info to GETPIDINFO prim. [FB6 change] - Changed all FLOAT precision to DOUBLE instead. [FB6 change]

+++ protomuck/src/debugger.c
@@ -387,7 +387,7 @@ static void
 push_arg(dbref player, struct frame *fr, const char *arg)
 {
     int num = 0, lflag = 0, sflag = 0;
-    float inum = 0.0;
+    double inum = 0.0;
 
 
     if (fr->argument.top >= STACK_SIZE) {
@@ -401,7 +401,7 @@ push_arg(dbref player, struct frame *fr, const char *arg)
 	anotify_nolisten(player, CSUCC "Integer pushed.", 1);
     } else if (ifloat(arg)) {
 	/* push a float */
-	inum = (float) atof(arg);
+	inum =  atof(arg);
 	push(fr->argument.st, &fr->argument.top, PROG_FLOAT, MIPSCAST & inum);
 	notify_nolisten(player, "Float pushed.", 1);
     } else if (*arg == NUMBER_TOKEN) {
